requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Finance
yfinance>=0.2.0

# Agno Framework (Agente Multi-Agente)
agno>=1.0.0
sqlalchemy>=2.0.0
chromadb>=0.4.0
google-genai>=0.1.0

# SDK e PDF
google-generativeai>=0.4.0
pypdf>=4.0.0
